Upgrade configurations for specific Windows users.
Updated the project configurations to ensure compatibility with Windows environments. This includes creating the .env file as a reminder for Windows users, modifying the .air.toml file to use `.exe` extension for binaries, and adjusting the package.json script for templ to remove unnecessary quotation marks. Additionally, upgraded the templ version to the latest release to ensure compatibility and take advantage of the latest features. This commit addresses issues related to configuration and ensures smooth execution on Windows platforms.